// src/app/shared/models/displayModels/basicModel.ts
export class BasicModel {
    id: number;
    folder_id: string;
    avatar_img: string;
    profile_img: string;
    name: string;
    address_1: string;
    address_2: string;
    address_3: string;
    city: string;
    summary: string;
    show_referees: boolean;
    show_repositories: boolean;
    avatar: any;
    profile: any;
}